backend/requirements.txt
asgiref==3.2.3
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==3.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
